Feature: multichain support

+++ rust/kadena/kadena-client/src/models/data/chainweb_response_meta_data_dto.rs
@@ -4,9 +4,9 @@ pub struct ChainwebResponseMetaDataDto {
     #[serde(rename = "blockHash", skip_serializing_if = "Option::is_none")]
     pub block_hash: Option<String>,
     #[serde(rename = "blockTime")]
-    pub block_time: f32,
+    pub block_time: u64,
     #[serde(rename = "blockHeight")]
-    pub block_height: f32,
+    pub block_height: u64,
     #[serde(rename = "prevBlockHash")]
     pub prev_block_hash: String,
     #[serde(rename = "publicMeta", skip_serializing_if = "Option::is_none")]
@@ -15,8 +15,8 @@ pub struct ChainwebResponseMetaDataDto {
 
 impl ChainwebResponseMetaDataDto {
     pub fn new(
-        block_time: f32,
-        block_height: f32,
+        block_time: u64,
+        block_height: u64,
         prev_block_hash: String,
     ) -> ChainwebResponseMetaDataDto {
         ChainwebResponseMetaDataDto {
