+++ rust/kadena/kadena-client/src/models/data/meta_data_dto.rs
@@ -2,13 +2,13 @@
 #[cfg_attr(test, derive(Default))]
 pub struct MetaDataDto {
     #[serde(rename = "creationTime")]
-    pub creation_time: f32,
+    pub creation_time: u64,
     #[serde(rename = "ttl")]
-    pub ttl: f32,
+    pub ttl: u64,
     #[serde(rename = "gasLimit")]
-    pub gas_limit: f32,
+    pub gas_limit: u64,
     #[serde(rename = "gasPrice")]
-    pub gas_price: f32,
+    pub gas_price: f64,
     #[serde(rename = "sender")]
     pub sender: String,
     #[serde(rename = "chainId")]
@@ -17,10 +17,10 @@ pub struct MetaDataDto {
 
 impl MetaDataDto {
     pub fn new(
-        creation_time: f32,
-        ttl: f32,
-        gas_limit: f32,
-        gas_price: f32,
+        creation_time: u64,
+        ttl: u64,
+        gas_limit: u64,
+        gas_price: f64,
         sender: String,
         chain_id: serde_json::Value,
     ) -> MetaDataDto {
